coords' type changed from int to float

+++ untitled-sdl-tank-game/Bullet.h
@@ -8,7 +8,7 @@ public:
 	int lifeCycle;
 	
 	Bullet();
-	Bullet(int x, int y);
+	Bullet(float x, float y);
 
 	~Bullet();
